# DOCUMENTATION/openapi.yaml
openapi: 3.0.4
info:
  title: Repetitor API
  description: REST API сервиса для изучения иностранных языков (Android / Web). Интервальное
    повторение, словарь с семантическим поиском (pgvector), генерация упражнений,
    чат-репетитор, TTS/STT и оценка произношения.
  contact:
    name: Repetitor
  version: v1
paths:
...
  /api/v1/teacher/groups:
    get:
      tags:
      - TeacherGroups
      summary: 'Возвращает группы учителя с количеством участников и курсов.

        Администратор получает все группы платформы.'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudyGroupResponse'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudyGroupResponse'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudyGroupResponse'
    post:
      tags:
      - TeacherGroups
      summary: 'Создаёт новую группу. Группа из одного участника — это персональные
        занятия.

        Возвращает 201 с данными созданной группы.'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyGroupRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/StudyGroupRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/StudyGroupRequest'
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/StudyGroupResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGroupResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/StudyGroupResponse'
...
  /api/v1/teacher/groups/{id}:
    get:
      tags:
      - TeacherGroups
      summary: 'Возвращает группу с составом, курсами и кодами-приглашениями.

        Возвращает 404, если группа не найдена, и 403, если она принадлежит другому
        учителю.'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/StudyGroupDetailResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGroupDetailResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/StudyGroupDetailResponse'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    patch:
      tags:
      - TeacherGroups
      summary: Изменяет название и описание группы. Переданные только те поля обновляются.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyGroupUpdateRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/StudyGroupUpdateRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/StudyGroupUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/StudyGroupResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGroupResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/StudyGroupResponse'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
      - TeacherGroups
      summary: 'Удаляет группу вместе с составом, курсами и приглашениями.

        Записи учеников на курсы и их прогресс сохраняются.'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/teacher/groups/{id}/members:
    post:
      tags:
      - TeacherGroups
      summary: 'Добавляет в группу зарегистрированного ученика по email и сразу выдаёт
        ему курсы группы.

        Повторное добавление не создаёт дублей. Возвращает 404, если ученик не зарегистрирован.'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddGroupMemberRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/AddGroupMemberRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AddGroupMemberRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/StudyGroupDetailResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGroupDetailResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/StudyGroupDetailResponse'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/teacher/groups/{id}/members/{userId}:
    delete:
      tags:
      - TeacherGroups
      summary: Исключает ученика из группы. Курсы, которые он уже открыл, остаются
        у него.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/StudyGroupDetailResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGroupDetailResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/StudyGroupDetailResponse'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/teacher/groups/{id}/invitations:
    post:
      tags:
      - TeacherGroups
      summary: 'Создаёт код-приглашение в группу: им может воспользоваться ученик,
        который ещё не зачислен.

        Срок действия задаётся в днях, лимит использований 0 означает без ограничений.'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateInvitationRequest'
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/GroupInvitationResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/GroupInvitationResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/GroupInvitationResponse'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    get:
      tags:
      - TeacherGroups
      summary: 'Возвращает коды-приглашения группы с их состоянием: активен, отозван
        или исчерпан.'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupInvitationResponse'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupInvitationResponse'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupInvitationResponse'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/teacher/groups/{id}/invitations/{invitationId}:
    delete:
      tags:
      - TeacherGroups
      summary: 'Отзывает код-приглашение: после отзыва ученик не сможет вступить по
        нему.'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: invitationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/teacher/groups/{id}/courses:
    post:
      tags:
      - TeacherGroups
      summary: 'Назначает группе курс и сразу зачисляет всех участников.

        Доступны собственные курсы учителя и опубликованные системные курсы;

        для остальных возвращается 400.'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignGroupCourseRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/AssignGroupCourseRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AssignGroupCourseRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/StudyGroupDetailResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGroupDetailResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/StudyGroupDetailResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/teacher/groups/{id}/courses/{courseId}:
    delete:
      tags:
      - TeacherGroups
      summary: Убирает курс из группы. Записи участников на курс и их прогресс сохраняются.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: courseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/StudyGroupDetailResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGroupDetailResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/StudyGroupDetailResponse'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  schemas:
...
    AddGroupMemberRequest:
      required:
      - email
      type: object
      properties:
        email:
          maxLength: 320
          minLength: 0
          type: string
          description: Email зарегистрированного ученика; сравнение регистронезависимое.
          format: email
      additionalProperties: false
      description: Добавление ученика в группу по его email.
...
    AssignGroupCourseRequest:
      required:
      - courseId
      type: object
      properties:
        courseId:
          type: string
          description: 'Курс: собственный курс учителя или опубликованный системный.'
          format: uuid
      additionalProperties: false
      description: Назначение курса группе.
...
    CreateInvitationRequest:
      type: object
      properties:
        expiresInDays:
          maximum: 365
          minimum: 1
          type: integer
          description: Срок действия кода в днях, от 1 до 365.
          format: int32
        maxUses:
          maximum: 1000
          minimum: 0
          type: integer
          description: Сколько учеников могут вступить по коду; 0 — без ограничения.
          format: int32
      additionalProperties: false
      description: Параметры нового кода-приглашения.
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: {}
...
    StudyGroupDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        teacherUserId:
          type: string
          format: uuid
        isPersonal:
          type: boolean
        createdAt:
          type: string
          format: date-time
        members:
          type: array
          items:
            $ref: '#/components/schemas/StudyGroupMemberResponse'
          nullable: true
        courses:
          type: array
          items:
            $ref: '#/components/schemas/StudyGroupCourseResponse'
          nullable: true
        invitations:
          type: array
          items:
            $ref: '#/components/schemas/GroupInvitationResponse'
          nullable: true
      additionalProperties: false
      description: Полное состояние группы для кабинета учителя.
...
    StudyGroupRequest:
      required:
      - name
      type: object
      properties:
        name:
          maxLength: 160
          minLength: 1
          type: string
          description: 'Название группы: для персональных занятий обычно имя ученика.'
        description:
          maxLength: 1000
          minLength: 0
          type: string
          description: 'Необязательное описание: цель, расписание, уровень.'
          nullable: true
      additionalProperties: false
      description: Запрос на создание или переименование учебной группы.
    StudyGroupResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        membersCount:
          type: integer
          format: int32
        coursesCount:
          type: integer
          format: int32
        isPersonal:
          type: boolean
        createdAt:
          type: string
          format: date-time
      additionalProperties: false
      description: Краткая карточка учебной группы в списке учителя.
    StudyGroupUpdateRequest:
      type: object
      properties:
        name:
          maxLength: 160
          minLength: 1
          type: string
          nullable: true
        description:
          maxLength: 1000
          minLength: 0
          type: string
          nullable: true
      additionalProperties: false
      description: 'Частичное обновление группы: null означает «поле не передано».'
...
  securitySchemes:
    Bearer:
      type: http
      description: Вставьте access-токен, полученный из /api/v1/auth/login
      scheme: bearer
      bearerFormat: JWT
security:
- {}
...